refactor: move state machine to state_machine.c
